# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GrailBabe API - collectibles portfolio tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [forum]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ForumPost" }
    post:
      operationId: createPost
      tags: [forum]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForumPostInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForumPost" }
  /posts/{id}:
    get:
      operationId: getPost
      tags: [forum]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForumPostDetail" }
  /posts/{id}/replies:
    post:
      operationId: createReply
      tags: [forum]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReplyInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForumReply" }
  /posts/{id}/react:
    post:
      operationId: reactPost
      tags: [forum]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReactionInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForumPost" }
...
  /community/posts:
    get:
      operationId: listCommunityPosts
      tags: [community]
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [hot, new, top] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CommunityPostSummary" }
    post:
      operationId: createCommunityPost
      tags: [community]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommunityPostInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommunityPostSummary" }
  /community/posts/{id}:
    get:
      operationId: getCommunityPost
      tags: [community]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommunityPostSummary" }
  /community/posts/{id}/comments:
    get:
      operationId: listCommunityComments
      tags: [community]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CommunityComment" }
    post:
      operationId: createCommunityComment
      tags: [community]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommunityCommentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommunityComment" }
  /community/posts/{id}/vote:
    post:
      operationId: voteCommunityPost
      tags: [community]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommunityVoteInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommunityPostSummary" }
...
components:
  schemas:
...
    ForumPost:
      type: object
      required: [id, userId, userName, title, body, replyCount, reactionCount, createdAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        userName: { type: string }
        userAvatar: { type: ["string", "null"] }
        title: { type: string }
        body: { type: string }
        category: { type: ["string", "null"] }
        replyCount: { type: integer }
        reactionCount: { type: integer }
        userReaction: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ForumPostDetail:
      type: object
      required: [post, replies]
      properties:
        post: { $ref: "#/components/schemas/ForumPost" }
        replies:
          type: array
          items: { $ref: "#/components/schemas/ForumReply" }
    ForumReply:
      type: object
      required: [id, postId, userId, userName, body, createdAt]
      properties:
        id: { type: integer }
        postId: { type: integer }
        userId: { type: string }
        userName: { type: string }
        userAvatar: { type: ["string", "null"] }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    ForumPostInput:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        category: { type: string }
    ReplyInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1 }
    ReactionInput:
      type: object
      required: [reaction]
      properties:
        reaction: { type: string, enum: [fire, heart, eyes, clap, none] }
...
    CommunityPostSummary:
      type: object
      required: [id, authorId, authorScreenname, authorName, category, title, body, isPinned, score, userVote, commentCount, createdAt]
      properties:
        id: { type: integer }
        authorId: { type: string }
        authorScreenname: { type: ["string", "null"] }
        authorName: { type: string }
        authorAvatar: { type: ["string", "null"] }
        category: { type: string }
        title: { type: string }
        body: { type: string }
        isPinned: { type: boolean }
        score: { type: integer }
        userVote: { type: integer, enum: [-1, 0, 1] }
        commentCount: { type: integer }
        createdAt: { type: string, format: date-time }
    CommunityPostInput:
      type: object
      required: [category, title, body]
      properties:
        category: { type: string, minLength: 1 }
        title: { type: string, minLength: 1, maxLength: 200 }
        body: { type: string, minLength: 1 }
    CommunityComment:
      type: object
      required: [id, postId, authorId, authorScreenname, authorName, body, createdAt]
      properties:
        id: { type: integer }
        postId: { type: integer }
        authorId: { type: string }
        authorScreenname: { type: ["string", "null"] }
        authorName: { type: string }
        authorAvatar: { type: ["string", "null"] }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    CommunityCommentInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1 }
    CommunityVoteInput:
      type: object
      required: [value]
      properties:
        value: { type: integer, enum: [-1, 1] }
...