# lib/api-zod/src/generated/types/communityComment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GrailBabe API - collectibles portfolio tracker
 * OpenAPI spec version: 0.1.0
 */

export interface CommunityComment {
  id: number;
  postId: number;
  authorId: string;
  /** @nullable */
  authorScreenname: string | null;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  body: string;
  createdAt: Date;
}