# lib/api-zod/src/generated/types/communityPostSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GrailBabe API - collectibles portfolio tracker
 * OpenAPI spec version: 0.1.0
 */
import type { CommunityPostSummaryUserVote } from './communityPostSummaryUserVote';

export interface CommunityPostSummary {
  id: number;
  authorId: string;
  /** @nullable */
  authorScreenname: string | null;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  category: string;
  title: string;
  body: string;
  isPinned: boolean;
  score: number;
  userVote: CommunityPostSummaryUserVote;
  commentCount: number;
  createdAt: Date;
}